[Modify] Already Vote / Closed Vote Condition
