+++ assets/css/dao.css
@@ -68,9 +68,14 @@
 .dao-quest-list-columns > ul > li > ul > li > h2 {width:100%; font-family: "Pretendard"; margin-top:10px; border-radius:6px; background:#fff; overflow:hidden; background-color: #e9e9e9;}
 .dao-quest-list-columns > ul > li > ul > li > h2 div {height:4px; background:#000; z-index: 999;}
 .dao-quest-list-columns > ul > li > .selectBtn {margin: 24px 0; font-size:14px; font-weight:300; font-family:'Pretendard'; color:#000;}
+.dao-quest-list-columns > ul > li > .selectBtn.vote-finish {position: relative;}
+.dao-quest-list-columns > ul > li > .selectBtn.vote-finish > div {filter: blur(4px);}
+.dao-quest-list-columns > ul > li > .selectBtn.vote-finish > .vote-alarm {position: absolute; top: 0; bottom: 0; left: 0; right: 0; background-color: rgba(0, 0, 0, 0.8); color: #fff; border-radius: 4px; display: flex; justify-content: center; align-items: center; font-weight: 600; font-size: 24px; filter: none;}
 .dao-quest-list-columns > ul > li > .selectBtn > div:nth-child(2) {display:flex; justify-content:space-between; align-items:center; margin-top:24px;}
-.dao-quest-list-columns > ul > li > .selectBtn > div > button {width:48%; height:56px; border-radius:4px; background-color:#EFEFEF; font-size:16px; color:#000; font-weight:300;}
+.dao-quest-list-columns > ul > li > .selectBtn > div > button {width:48%; height:56px; border-radius:4px; background-color:#EFEFEF; font-size:16px; color:#000; font-weight:300; transition: all 0.3s;}
+.dao-quest-list-columns > ul > li > .selectBtn > div > button.answerBtn {width:100%;}
 .dao-quest-list-columns > ul > li > .selectBtn > div > button:hover {background-color:#000; color:#fff;}
+/* .dao-quest-list-columns > ul > li > .selectBtn > div > button:disabled {position: relative; overflow: hidden; background-color: rgba(0,0,0,0.7); color: #fff;} */
 .dao-quest-list-columns > ul > li > .selectBtn > .adminConfirmBtn {width: 100%; height: 56px; border-radius:4px; background-color:#EFEFEF; font-size:16px; color:#000; font-weight:300; margin-top: 8px;}
 .dao-quest-list-columns > ul > li > .selectBtn > .adminConfirmBtn:hover {background-color:#000; color:#fff;}
